requirements.txt
python-dotenv==0.17.1
bcrypt==3.2.0
cffi==1.14.5
click==7.1.2
dateparser==1.0.0
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pycparser==2.20
PyJWT==2.1.0
python-dateutil==2.8.1
pytz==2021.1
regex==2021.4.4
six==1.16.0
tzlocal==2.1
Werkzeug==1.0.1
